+++ Controller/Controller.hpp
@@ -15,7 +15,7 @@ class Controller
         Controller();
 
         //Returns a tuple of title, description, length, returns {"", "", 0.0f} if Id is invalid
-        tuple<string, string, float> GetDocument(string _documentId);
+        tuple<string, string, double> GetDocument(string _documentId);
         //Returns a tuple of name, description, returns {"", ""} if Id is invalid
         tuple<string, string> GetContainer(string _containerId);
         //Returns a set of the documents in container Ids
@@ -24,7 +24,7 @@ class Controller
         unordered_set<string> GetContainers(string _documentId);
 
         //Returns 1 if there's already a doc or a container with the same Id, 0 else
-        int AddDocument(string _id, string _title, string _description, float _length);
+        int AddDocument(string _id, string _title, string _description, double _length);
         //Returns 1 if there isn't a doc with this Id, 0 else
         int RemoveDocument(string _id);
         //Returns 1 if there's already a container or a doc with the same Id, 0 else
